move enum into config due to import pattern
